Added comment for constructor method
A constructor method allows for an alternative way of declaring,
initializing, and accessing the variables.

Originally didn't do it this way because I was unsure if this would be
considered as "global variables" since they are instances of the class.
Adding it for the possibility of changing to it later since it can make
the code a little bit cleaner since global keyword is not required to
access the variables.