Updated comment handling
Clears current token if a comment is identified in next_token().
Previously, a comment would be identified if the token was a double
slash but a comment should not be stored as a token, including the
double token.

In read_file(), to match the comment handing of next_token(), setting
token type and printing the token is only done if the token is nonempty.